+++ src/types.ts
@@ -22,10 +22,13 @@ export type CharacterDataList = { [key: string]: CharacterData };
 export type Dictionary = {
   [key: string]: {
     meaning: string;
-    definition: string;
     exampleSentence: string;
     exampleMeaning: string;
   };
+};
+
+export type DefinitionMini = {
+  [key: string]: string;
 };
 
 export type Primitives = {